refactor: share makeIconBtn
